Add Robolectric test support on :pager
