+++ pager/build.gradle
@@ -71,6 +71,17 @@ android {
         }
         animationsDisabled true
     }
+
+    sourceSets {
+        test {
+            java.srcDirs += 'src/sharedTest/kotlin'
+            res.srcDirs += 'src/sharedTest/res'
+        }
+        androidTest {
+            java.srcDirs += 'src/sharedTest/kotlin'
+            res.srcDirs += 'src/sharedTest/res'
+        }
+    }
 }
 
 dependencies {
@@ -81,13 +92,25 @@ dependencies {
     // Test dependencies
     // ======================
 
+    androidTestImplementation project(':internal-testutils')
+    testImplementation project(':internal-testutils')
+
     androidTestImplementation libs.junit
+    testImplementation libs.junit
+
     androidTestImplementation libs.truth
+    testImplementation libs.truth
 
     androidTestImplementation libs.compose.ui.test.junit4
+    testImplementation libs.compose.ui.test.junit4
+
     androidTestImplementation libs.compose.ui.test.manifest
-    androidTestImplementation libs.compose.foundation.foundation
+    testImplementation libs.compose.ui.test.manifest
+
     androidTestImplementation libs.androidx.test.runner
+    testImplementation libs.androidx.test.runner
+
+    testImplementation libs.robolectric
 }
 
 apply plugin: "com.vanniktech.maven.publish"
