+++ pager/src/androidTest/kotlin/com/google/accompanist/pager/InstrumentedHorizontalPagerTest.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2021 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.accompanist.pager
+
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.unit.LayoutDirection
+import org.junit.runner.RunWith
+import org.junit.runners.Parameterized
+
+@RunWith(Parameterized::class)
+class InstrumentedHorizontalPagerTest(
+    itemWidthFraction: Float,
+    horizontalAlignment: Alignment.Horizontal,
+    itemSpacingDp: Int,
+    offscreenLimit: Int,
+    layoutDirection: LayoutDirection,
+    reverseLayout: Boolean,
+    infiniteLoop: Boolean,
+) : BaseHorizontalPagerTest(
+    itemWidthFraction,
+    horizontalAlignment,
+    itemSpacingDp,
+    offscreenLimit,
+    layoutDirection,
+    reverseLayout,
+    infiniteLoop,
+) {
+    companion object {
+        @JvmStatic
+        @Parameterized.Parameters
+        fun data() = params()
+    }
+}
