+++ pager/src/sharedTest/kotlin/com/google/accompanist/pager/BaseHorizontalPagerTest.kt
@@ -29,124 +29,122 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.SemanticsNodeInteraction
 import androidx.compose.ui.test.assertHeightIsAtLeast
 import androidx.compose.ui.test.assertLeftPositionInRootIsEqualTo
+import androidx.compose.ui.test.assertTopPositionInRootIsEqualTo
 import androidx.compose.ui.test.assertWidthIsEqualTo
 import androidx.compose.ui.test.getUnclippedBoundsInRoot
 import androidx.compose.ui.test.onRoot
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.height
 import androidx.compose.ui.unit.width
-import androidx.test.filters.LargeTest
-import com.google.common.truth.Truth.assertThat
-import org.junit.runner.RunWith
-import org.junit.runners.Parameterized
+import com.google.accompanist.imageloading.test.combineWithParameters
+import com.google.accompanist.imageloading.test.parameterizedParams
 
 @OptIn(ExperimentalPagerApi::class) // Pager is currently experimental
-@LargeTest
-@RunWith(Parameterized::class)
-class VerticalPagerTest(
-    private val verticalAlignment: Alignment.Vertical,
+abstract class BaseHorizontalPagerTest(
+    private val itemWidthFraction: Float,
+    private val horizontalAlignment: Alignment.Horizontal,
     // We don't use the Dp type due to https://youtrack.jetbrains.com/issue/KT-35523
     private val itemSpacingDp: Int,
     override val offscreenLimit: Int,
+    private val layoutDirection: LayoutDirection,
     private val reverseLayout: Boolean,
     override val infiniteLoop: Boolean,
 ) : PagerTest() {
-    companion object {
+    protected companion object {
         @JvmStatic
-        @Parameterized.Parameters
-        fun data(): Collection<Array<Any>> = dataset(false) + dataset(true)
-
-        private fun dataset(looping: Boolean): Collection<Array<Any>> = listOf(
-            // verticalAlignment, itemSpacingDp, offscreenLimit, reverseLayout, infiniteLoop
-
-            // Test typical full-width items
-            arrayOf(Alignment.CenterVertically, 0, 2, false, looping),
-            arrayOf(Alignment.Top, 0, 2, false, looping),
-            arrayOf(Alignment.Bottom, 0, 2, false, looping),
-
-            // Full-width items with spacing
-            arrayOf(Alignment.CenterVertically, 4, 2, false, looping),
-            arrayOf(Alignment.Top, 4, 2, false, looping),
-            arrayOf(Alignment.Bottom, 4, 2, false, looping),
-
-            // Full-width items with reverseLayout = true
-            arrayOf(Alignment.CenterVertically, 0, 2, true, looping),
-            arrayOf(Alignment.Top, 0, 2, true, looping),
-            arrayOf(Alignment.Bottom, 0, 2, true, looping),
-
-            // Test an increased offscreenLimit
-            arrayOf(Alignment.CenterVertically, 0, 4, false, looping),
-        )
+        protected fun params(): Collection<Array<Any>> = parameterizedParams()
+            // itemWidthFraction
+            .combineWithParameters(1f, 0.8f)
+            // horizontalAlignment
+            .combineWithParameters(Alignment.CenterHorizontally, Alignment.Start, Alignment.End)
+            // itemSpacing
+            .combineWithParameters(0, 4)
+            // offscreenLimit
+            .combineWithParameters(1, 2)
+            // layoutDirection
+            .combineWithParameters(LayoutDirection.Ltr, LayoutDirection.Rtl)
+            // reverseLayout
+            .combineWithParameters(true, false)
+            // looping
+            .combineWithParameters(true, false)
     }
+
+    /**
+     * Returns the expected resolved layout direction for pages
+     */
+    private val reverseDirection: Boolean
+        get() = if (layoutDirection == LayoutDirection.Rtl) !reverseLayout else reverseLayout
 
     override fun SemanticsNodeInteraction.swipeAcrossCenter(
         distancePercentage: Float,
         velocity: Float,
     ): SemanticsNodeInteraction = swipeAcrossCenterWithVelocity(
-        distancePercentageY = if (reverseLayout) -distancePercentage else distancePercentage,
+        distancePercentageX = if (reverseDirection) -distancePercentage else distancePercentage,
         velocity = velocity,
     )
 
     override fun SemanticsNodeInteraction.assertLaidOutItemPosition(
         page: Int,
-        currentPage: Int
+        currentPage: Int,
     ): SemanticsNodeInteraction {
         val rootBounds = composeTestRule.onRoot().getUnclippedBoundsInRoot()
-        val expectedItemSize = rootBounds.width
+        val expectedItemSize = rootBounds.width * itemWidthFraction
 
         // The expected coordinates. This uses the implicit fact that VerticalPager by
         // use Alignment.CenterVertically by default, and that we're using items
         // with an aspect ratio of 1:1
-        val expectedLeft = (rootBounds.width - expectedItemSize) / 2
-        val expectedFirstItemTop = when (verticalAlignment) {
-            Alignment.Top -> 0.dp
-            Alignment.Bottom -> rootBounds.height - expectedItemSize
-            else /* Alignment.CenterVertically */ -> (rootBounds.height - expectedItemSize) / 2
+        val expectedTop = (rootBounds.height - expectedItemSize) / 2
+        val expectedFirstItemLeft = when (horizontalAlignment) {
+            Alignment.Start -> {
+                when (layoutDirection) {
+                    LayoutDirection.Ltr -> 0.dp
+                    else -> rootBounds.width - expectedItemSize
+                }
+            }
+            Alignment.End -> {
+                when (layoutDirection) {
+                    LayoutDirection.Ltr -> rootBounds.width - expectedItemSize
+                    else -> 0.dp
+                }
+            }
+            else /* Alignment.CenterVertically */ -> (rootBounds.width - expectedItemSize) / 2
         }
 
         return assertWidthIsEqualTo(expectedItemSize)
             .assertHeightIsAtLeast(expectedItemSize)
-            .assertLeftPositionInRootIsEqualTo(expectedLeft)
+            .assertTopPositionInRootIsEqualTo(expectedTop)
             .run {
                 val pageDelta = ((expectedItemSize + itemSpacingDp.dp) * (page - currentPage))
-                // Not sure why, but there's a rounding error somewhere.
-                // TODO: reduce this tolerance and work out where the issue is
-                if (reverseLayout) {
-                    assertTopPositionInRootIsEqualTo(
-                        expectedTop = expectedFirstItemTop - pageDelta,
-                        tolerance = 1.dp
-                    )
+                if (reverseDirection) {
+                    assertLeftPositionInRootIsEqualTo(expectedFirstItemLeft - pageDelta)
                 } else {
-                    assertTopPositionInRootIsEqualTo(
-                        expectedTop = expectedFirstItemTop + pageDelta,
-                        tolerance = 1.dp
-                    )
+                    assertLeftPositionInRootIsEqualTo(expectedFirstItemLeft + pageDelta)
                 }
             }
     }
 
-    override fun setPagerContent(pageCount: Int): PagerState {
+    override fun setPagerContent(
+        pageCount: Int,
+    ): PagerState {
         val pagerState = PagerState(
             pageCount = pageCount,
             offscreenLimit = offscreenLimit,
             infiniteLoop = infiniteLoop,
         ).apply { testing = true }
-        // Stick to LTR for vertical tests
-        composeTestRule.setContent(LayoutDirection.Ltr) {
+        composeTestRule.setContent(layoutDirection) {
             applierScope = rememberCoroutineScope()
 
-            VerticalPager(
+            HorizontalPager(
                 state = pagerState,
                 itemSpacing = itemSpacingDp.dp,
                 reverseLayout = reverseLayout,
-                verticalAlignment = verticalAlignment,
+                horizontalAlignment = horizontalAlignment,
                 modifier = Modifier.fillMaxSize()
             ) { page ->
                 Box(
                     modifier = Modifier
-                        .fillMaxWidth()
+                        .fillMaxWidth(itemWidthFraction)
                         .aspectRatio(1f)
                         .background(randomColor())
                         .testTag(page.toString())
@@ -160,17 +158,4 @@ class VerticalPagerTest(
         }
         return pagerState
     }
-}
-
-/**
- * A version of `assertTopPositionInRootIsEqualTo` which allows setting of the tolerance.
- */
-private fun SemanticsNodeInteraction.assertTopPositionInRootIsEqualTo(
-    expectedTop: Dp,
-    tolerance: Dp = 0.5.dp
-): SemanticsNodeInteraction {
-    assertThat(getUnclippedBoundsInRoot().top.value)
-        .isWithin(tolerance.value)
-        .of(expectedTop.value)
-    return this
 }
